Move sklearn module out of ml_estimators
As sklearn has grown, it makes sense to move it up a level to be
a first-level submodule.
