+++ src/tests/fixtures/fake_estimator.py
@@ -5,7 +5,7 @@ from sklearn.compose import ColumnTransformer
 from sklearn.preprocessing import OneHotEncoder
 
 from machine_learning.ml_estimators import BaseMLEstimator
-from machine_learning.ml_estimators.sklearn import ColumnDropper
+from machine_learning.sklearn import ColumnDropper
 from machine_learning.ml_data import MLData
 from machine_learning.settings import (
     TEAM_NAMES,
